113660: Update phone number utility function to be shared between all apps
