+++ src/applications/appeals/995/config/submit-transformer.js
@@ -6,11 +6,13 @@ import {
   getEmail,
   getEvidence,
   getForm4142,
-  getPhone,
   getMstData,
 } from '../utils/submit';
-
-import { addIncludedIssues, getTimeZone } from '../../shared/utils/submit';
+import {
+  getPhone,
+  addIncludedIssues,
+  getTimeZone,
+} from '../../shared/utils/submit';
 
 export function transform(formConfig, form) {
   // https://developer.va.gov/explore/appeals/docs/decision_reviews?version=current
@@ -31,7 +33,7 @@ export function transform(formConfig, form) {
       veteran: {
         timezone: getTimeZone(),
         address: getAddress(formData),
-        phone: getPhone(formData),
+        phone: getPhone(formData, true),
         email: getEmail(formData),
       },
       ...getEvidence(formData),
