+++ src/applications/appeals/995/utils/submit/index.js
@@ -2,7 +2,7 @@ import { CLAIMANT_TYPES } from '../../constants';
 import { MAX_LENGTH } from '../../../shared/constants';
 
 import { getHomeless } from './homeless';
-import { getAddress, getPhone, getEmail } from './veteran';
+import { getAddress, getEmail } from './veteran';
 import {
   getTreatmentDate,
   hasDuplicateLocation,
@@ -45,7 +45,6 @@ export const getClaimantData = ({
 export {
   getHomeless,
   getAddress,
-  getPhone,
   getEmail,
   getTreatmentDate,
   hasDuplicateLocation,
